Split main views of application into 2 separate components. :spaghetti:
