+++ .eslintcache
@@ -1,3 +1,1 @@
-
-[{"D:\\webDevelopment\\bootcamp\\project6\\caitlin-inga-victor-saeed-project-six\\src\\index.js":"1","D:\\webDevelopment\\bootcamp\\project6\\caitlin-inga-victor-saeed-project-six\\src\\reportWebVitals.js":"2","D:\\webDevelopment\\bootcamp\\project6\\caitlin-inga-victor-saeed-project-six\\src\\App.js":"3","D:\\webDevelopment\\bootcamp\\project6\\caitlin-inga-victor-saeed-project-six\\src\\firebase.js":"4"},{"size":523,"mtime":1607364186738,"results":"5","hashOfConfig":"6"},{"size":375,"mtime":1607290520418,"results":"7","hashOfConfig":"6"},{"size":6113,"mtime":1607371682018,"results":"8","hashOfConfig":"6"},{"size":545,"mtime":1607296297639,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1jdv8np",{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},{"filePath":"15","messages":"16","errorCount":0,"warningCount":3,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},"D:\\webDevelopment\\bootcamp\\project6\\caitlin-inga-victor-saeed-project-six\\src\\index.js",[],"D:\\webDevelopment\\bootcamp\\project6\\caitlin-inga-victor-saeed-project-six\\src\\reportWebVitals.js",[],["19","20"],"D:\\webDevelopment\\bootcamp\\project6\\caitlin-inga-victor-saeed-project-six\\src\\App.js",["21","22","23"],"D:\\webDevelopment\\bootcamp\\project6\\caitlin-inga-victor-saeed-project-six\\src\\firebase.js",[],{"ruleId":"24","replacedBy":"25"},{"ruleId":"26","replacedBy":"27"},{"ruleId":"28","severity":1,"message":"29","line":2,"column":27,"nodeType":"30","messageId":"31","endLine":2,"endColumn":33},{"ruleId":"28","severity":1,"message":"32","line":2,"column":35,"nodeType":"30","messageId":"31","endLine":2,"endColumn":40},{"ruleId":"28","severity":1,"message":"33","line":2,"column":42,"nodeType":"30","messageId":"31","endLine":2,"endColumn":46},"no-native-reassign",["34"],"no-negated-in-lhs",["35"],"no-unused-vars","'Router' is defined but never used.","Identifier","unusedVar","'Route' is defined but never used.","'Link' is defined but never used.","no-global-assign","no-unsafe-negation"]
-
+[{"C:\\Users\\victo\\Desktop\\General Directory\\Web Development\\Juno College Fulltime Cohort30\\week7-nov30\\Project 6 - Not So Social Event Planner\\caitlin-inga-victor-saeed-project-six\\src\\index.js":"1","C:\\Users\\victo\\Desktop\\General Directory\\Web Development\\Juno College Fulltime Cohort30\\week7-nov30\\Project 6 - Not So Social Event Planner\\caitlin-inga-victor-saeed-project-six\\src\\reportWebVitals.js":"2","C:\\Users\\victo\\Desktop\\General Directory\\Web Development\\Juno College Fulltime Cohort30\\week7-nov30\\Project 6 - Not So Social Event Planner\\caitlin-inga-victor-saeed-project-six\\src\\App.js":"3","C:\\Users\\victo\\Desktop\\General Directory\\Web Development\\Juno College Fulltime Cohort30\\week7-nov30\\Project 6 - Not So Social Event Planner\\caitlin-inga-victor-saeed-project-six\\src\\firebase.js":"4","C:\\Users\\victo\\Desktop\\General Directory\\Web Development\\Juno College Fulltime Cohort30\\week7-nov30\\Project 6 - Not So Social Event Planner\\caitlin-inga-victor-saeed-project-six\\src\\ResultsPage.js":"5","C:\\Users\\victo\\Desktop\\General Directory\\Web Development\\Juno College Fulltime Cohort30\\week7-nov30\\Project 6 - Not So Social Event Planner\\caitlin-inga-victor-saeed-project-six\\src\\SocialEvents.js":"6"},{"size":523,"mtime":1607373376298,"results":"7","hashOfConfig":"8"},{"size":375,"mtime":1607285315027,"results":"9","hashOfConfig":"8"},{"size":3348,"mtime":1607378040338,"results":"10","hashOfConfig":"8"},{"size":545,"mtime":1607373376296,"results":"11","hashOfConfig":"8"},{"size":2145,"mtime":1607378091985,"results":"12","hashOfConfig":"8"},{"size":2859,"mtime":1607378124479,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},"1jz52s3",{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"19","messages":"20","errorCount":0,"warningCount":4,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"23","messages":"24","errorCount":0,"warningCount":2,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"25","messages":"26","errorCount":0,"warningCount":3,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"C:\\Users\\victo\\Desktop\\General Directory\\Web Development\\Juno College Fulltime Cohort30\\week7-nov30\\Project 6 - Not So Social Event Planner\\caitlin-inga-victor-saeed-project-six\\src\\index.js",[],["27","28"],"C:\\Users\\victo\\Desktop\\General Directory\\Web Development\\Juno College Fulltime Cohort30\\week7-nov30\\Project 6 - Not So Social Event Planner\\caitlin-inga-victor-saeed-project-six\\src\\reportWebVitals.js",[],"C:\\Users\\victo\\Desktop\\General Directory\\Web Development\\Juno College Fulltime Cohort30\\week7-nov30\\Project 6 - Not So Social Event Planner\\caitlin-inga-victor-saeed-project-six\\src\\App.js",["29","30","31","32"],"C:\\Users\\victo\\Desktop\\General Directory\\Web Development\\Juno College Fulltime Cohort30\\week7-nov30\\Project 6 - Not So Social Event Planner\\caitlin-inga-victor-saeed-project-six\\src\\firebase.js",[],"C:\\Users\\victo\\Desktop\\General Directory\\Web Development\\Juno College Fulltime Cohort30\\week7-nov30\\Project 6 - Not So Social Event Planner\\caitlin-inga-victor-saeed-project-six\\src\\ResultsPage.js",["33","34"],"C:\\Users\\victo\\Desktop\\General Directory\\Web Development\\Juno College Fulltime Cohort30\\week7-nov30\\Project 6 - Not So Social Event Planner\\caitlin-inga-victor-saeed-project-six\\src\\SocialEvents.js",["35","36","37"],{"ruleId":"38","replacedBy":"39"},{"ruleId":"40","replacedBy":"41"},{"ruleId":"42","severity":1,"message":"43","line":2,"column":27,"nodeType":"44","messageId":"45","endLine":2,"endColumn":33},{"ruleId":"42","severity":1,"message":"46","line":2,"column":35,"nodeType":"44","messageId":"45","endLine":2,"endColumn":40},{"ruleId":"42","severity":1,"message":"47","line":2,"column":42,"nodeType":"44","messageId":"45","endLine":2,"endColumn":46},{"ruleId":"48","severity":1,"message":"49","line":91,"column":3,"nodeType":"50","messageId":"51","endLine":94,"endColumn":4},{"ruleId":"42","severity":1,"message":"46","line":2,"column":35,"nodeType":"44","messageId":"45","endLine":2,"endColumn":40},{"ruleId":"42","severity":1,"message":"47","line":2,"column":42,"nodeType":"44","messageId":"45","endLine":2,"endColumn":46},{"ruleId":"42","severity":1,"message":"43","line":2,"column":27,"nodeType":"44","messageId":"45","endLine":2,"endColumn":33},{"ruleId":"42","severity":1,"message":"46","line":2,"column":35,"nodeType":"44","messageId":"45","endLine":2,"endColumn":40},{"ruleId":"42","severity":1,"message":"47","line":2,"column":42,"nodeType":"44","messageId":"45","endLine":2,"endColumn":46},"no-native-reassign",["52"],"no-negated-in-lhs",["53"],"no-unused-vars","'Router' is defined but never used.","Identifier","unusedVar","'Route' is defined but never used.","'Link' is defined but never used.","no-useless-constructor","Useless constructor.","MethodDefinition","noUselessConstructor","no-global-assign","no-unsafe-negation"]
